+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.1.0]
+
+### Changed
+
+- Removed `RustyAPIConfig`; The new `NostrAuthApiConfig` replaces it partially.
+
+### Added
+
+- New `NostrAuthApiConfig` for default nostr authentication
+
 ## [0.0.3]
 
 ### Changed